kokoro: run tests with Trampoline (#451)
This also moves the Kokoro configuration into the testing directory.

system_tests.sh is based on testing/cloudbuild/run.bash with some
modifications to make it work with Kokoro. In particular, the service
account key is stored in KOKORO_KEYSTORE_DIR and the repo is checked out
to a `github` subdirectory.

Builds are not cached and every project uses the same source image (from
the golang-samples-tests project). This could be improved in the future,
but I'm not sure the best way to dynamically use different source images
in Trampoline (which starts the image).