all: remove legacy health check handlers (#1442)
Legacy health checks migrated to split health checks, which are now the
default. Split health checks are not sent to apps by default, so these
handlers are not working as intended.

None of the app.yaml files have health checks configured, so I chose not
to add new split handlers & configuration. I'm open to adding them
if we decide we should.

Fixes #977.