test(spanner): increase timeout and cleanup test (#1771)
Inserting 800 rows as a single Batch DML statement can take a long time, and the
transaction can be aborted one or more times. The deadline used for such an operation
therefore needs to be (very) long. It is now increased to 10 minutes.

Furthermore, the test case has been cleaned up a little bit:
* The multiple deferred functions for cleaning up have been combined into one.
* The database name is no longer fixed, allowing multiple instances of this test case
  to be executed at the same time.
* The generated player id's are now guaranteed to be unique. The previous version
  generated random id values that could overlap (although the probability was extremely
  low).

Fixes #1726